edit: 오픈 에이아이 서비스 인터페이스
