+++ back/src/main/java/org/ozea/common/controller/HomeController.java
@@ -16,17 +16,21 @@ import java.util.Map;
 @Slf4j
 public class HomeController {
     private final UserMapper userMapper;
+
     public HomeController(UserMapper userMapper) {
         this.userMapper = userMapper;
     }
+
     @GetMapping("/")
     public String home() {
         return "index";
     }
+
     @GetMapping("/signup")
     public String showSignupPage() {
         return "signup";
     }
+
     @PostMapping("/signup")
     public String processSignup(HttpServletRequest request) {
         String name = request.getParameter("name");
@@ -48,18 +52,14 @@ public class HomeController {
         request.getSession().setAttribute("email", email);
         return "redirect:/additional-info";
     }
+
     @GetMapping("/local-login")
     public String showLocalLoginPage() {
         return "local-login";
     }
+
     @GetMapping("/main")
     public String mainPage() {
         return "main";
-    }
-
-    @GetMapping("/__ping/date")
-    @ResponseBody
-    public Map<String, Object> ping() {
-        return Map.of("today", java.time.LocalDate.of(2025, 8, 9));
     }
 }
